requirements.txt
Django==3.2
mysqlclient==2.0.3
django-environ==0.4.5
aiogram==2.13
python-dotenv==0.17.0
SQLAlchemy==1.4.12
aiomysql==0.0.21
aioredis==1.3.1
